api/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
crewai[anthropic]>=0.70.0
anthropic>=0.40.0
requests>=2.31.0
python-multipart>=0.0.6
crewai-tools>=0.8.0
pyyaml>=6.0
asyncpg>=0.29.0